abstracted Shader

+++ Hazel/src/Hazel/Renderer/IndexBuffer.h
@@ -10,9 +10,9 @@ namespace hazel
 		virtual void bind() const = 0;
 		virtual void unbind() const = 0;
 
-		virtual size_t getCount() const = 0;
+		virtual uint32_t getCount() const = 0;
 
-		static IndexBuffer* create(uint32_t* vertices, size_t count);
+		static IndexBuffer* create(uint32_t* vertices, uint32_t count);
 	};
 }
