+++ Hazel/src/Platform/OpenGL/OpenGLIndexBuffer.h
@@ -7,13 +7,13 @@ namespace hazel
 	class OpenGLIndexBuffer : public IndexBuffer
 	{
 	public:
-		OpenGLIndexBuffer(uint32_t* vertices, size_t count);
+		OpenGLIndexBuffer(uint32_t* vertices, uint32_t count);
 		virtual ~OpenGLIndexBuffer();
 
 		void bind() const override;
 		void unbind() const override;
 
-		size_t getCount() const override { return count; };
+		uint32_t getCount() const override { return count; };
 
 	private:
 		uint32_t rendererID;
